Get objects from the RATO ArcGIS Enterprise environment via the API

This function retrieves objects from the RATO ArcGIS environment based on the object_id. It is capable of retrieving a large amount of objects in a single function call. As to minimize errors it will fetch batch size number of objects per request, and will operate in parallel: for a maximum of 5000 requests per 60 seconds, having at most 10 requests open at a time.

Usage

get_objects(
  object_ids = list_object_ids(),
  token = get_token(),
  batch_size = 50
)

Arguments

object_ids

Vector of object_ids, will be converted to character.

token

Access token, by default generated by get_token()

batch_size

Number of objects to request per API call, default is 50. Setting this to a lower number will result in more API calls, but will also reduce the risk of timeouts or errors when requesting large datasets. When set to a higher number, the API may timeout or return an error if the dataset is too large.

Value

tibble of requested objects.