Request an access token from the GIS API services

The credentials are stored in your .Renviron file under RATO_USER and RATO_PWD. If you haven't stored them there already the function will prompt you for them and store them in these environmental variables. To reset any stored values, pass emtpy strings "" to either the username or password arguments.

Usage

get_token(
  username = Sys.getenv("RATO_USER"),
  password = Sys.getenv("RATO_PWD")
)

Arguments

username

ArcGIS Enterprise username

password

ArcGIS Enterprise password

Value

Character. An access token for future API calls.

Details

By default tokens expire every 5 minutes, and are cached until they expire. You can set a different expiry duration by changing the ratatouille.rato_expires_minutes option with options()